requirements/developer.txt
# Install the main set of requirements
-r ./base.txt

# Additional requirements useful for developers of Boilerplate Webapp

# Django debugging
django-debug-toolbar
